requirements.txt
python-telegram-bot==20.7
openai==1.55.3
httpx==0.27.2
anyio==4.6.2
pydantic==2.9.2
requests==2.32.3
typing-extensions==4.12.2
